move components in guui and move out frontend specific dependencies
