+++ frontend/lib/assets.js
